+++ pages/blog/first-post.js
@@ -1,9 +1,0 @@
-export default function FirstPost() {
-  return (
-    <div className='post'>
-      <h2>First Post</h2>
-      <p>We're building this page.</p>
-      <p>Stay tuned!</p>
-    </div>
-  );
-}