+++ pages/blog/index.js
@@ -1,25 +1,45 @@
+import { useState } from 'react';
 import BlogPostsContainer from '@/components/blog/BlogPostsContainer';
-import Container from '@/components/containers/Container';
 import SearchBar from '@/components/blog/SearchBar';
-import styles from '@/styles/Blog.module.scss';
 import Title from '@/components/snippets/Title';
+import styles from '@/styles/Blog.module.scss';
+import { blogRevalidate } from '@/utils/config';
+import { tagToHeading } from '@/utils/blogCategories';
+import { blogSearch } from '@/utils/search';
 
 export default function Blog({ posts }) {
-  const latestPosts = posts.slice(0, 3);
-  const nextJsPosts = posts.filter(post => post.tagList.includes('nextjs'));
-  const typescriptPosts = posts.filter(post =>
-    post.tagList.includes('typescript')
-  );
+  const [searchTerm, setSearchTerm] = useState('');
+
+  const filteredData = {
+    posts: posts.slice(0, 3),
+  };
+
+  if (searchTerm) {
+    const filteredPosts = blogSearch(posts, searchTerm);
+    filteredData.posts = filteredPosts;
+    filteredData.heading = `${
+      filteredPosts.length === 0 ? 'no' : filteredPosts.length
+    } search ${
+      filteredPosts.length > 1 ? 'results' : 'result'
+    } for '${searchTerm}'`;
+    filteredData.viewall = false;
+  }
 
   return (
     <>
       <div className={styles.blogSearch}>
-        <Title customClass='blogTitle' title='Latest Posts' />
-        {/* <SearchBar /> */}
+        <Title customClass='blogTitle' title={!searchTerm && 'Latest Posts'} />
+        <SearchBar setSearchTerm={setSearchTerm} />
       </div>
-      <BlogPostsContainer posts={latestPosts} />
-      <BlogPostsContainer posts={nextJsPosts} tag='nextjs' />
-      <BlogPostsContainer posts={typescriptPosts} tag='typescript' />
+      <BlogPostsContainer {...filteredData} />
+      {!searchTerm &&
+        Object.keys(tagToHeading).map(tag => (
+          <BlogPostsContainer
+            key={tag}
+            posts={posts.filter(post => post.tagList.includes(tag))}
+            tag={tag}
+          />
+        ))}
     </>
   );
 }
@@ -40,5 +60,6 @@ export async function getStaticProps() {
         tagList: post.tag_list,
       })),
     },
+    revalidate: blogRevalidate,
   };
 }