+++ CHANGELOG.md
@@ -84,6 +84,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Blog page which pulls content from dev.to
+- Search functionality for blog posts
 
 ### Fixed